requirements.txt
langchain==0.1.16
langchain-groq==0.1.3
pandas
numpy
python-dotenv
